fix: Unify growth analysis logic for consistency
I refactored the growth analysis logic into a single, reusable utility function to resolve inconsistencies between the growth chart and health analysis pages.

- I created `client/src/utils/growth-analyzer.js` to house the accurate, interpolation-based analysis logic.
- I updated `GrowthChartPage.js` to use the new utility, fixing a bug where it used inaccurate calculations and was hardcoded for the wrong gender.
- I simplified the logic in `HealthAnalysisPage.js` to use the same utility.
